backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
awsebcli==3.25
bcrypt==4.3.0
blessed==1.21.0
boto3==1.39.8
botocore==1.39.8
cement==2.10.14
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
fabric==3.2.2
fastapi==0.116.1
fastapi-utils==0.8.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
invoke==2.2.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-tavily==0.2.7
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.7
multidict==6.6.3
openai==1.97.0
orjson==3.11.0
ormsgpack==1.10.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
propcache==0.3.2
psutil==5.9.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
termcolor==2.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
